requirements.txt
flask==3.0.3
groq==0.9.0
pypdf==4.3.1
python-dotenv==1.0.1
bcrypt==4.1.3
httpx==0.28.1
pymupdf
rank-bm25==0.2.1